fitbit_signage_api/requirements.txt
Django==3.2
mysqlclient
djangorestframework
django-filter
python-dotenv
djangorestframework-camel-case
fitbit